[NIAK-62] Moved RouteProvider
