+++ rezerwacje-front/src/common/index.js
@@ -1,1 +1,1 @@
-export { default } from './RouteProvider';
+export { default } from '../redux/routes/RouteProvider';
